requirements.txt
redis
requests
aioredis
aiohttp 
bitstring
aiofcm
python-dotenv
pylint
uvloop
python-socketio
websockets
python-rapidjson